# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MiAu API REST",
    "description": "\n## MiAu — API REST de Gestão Petshop\n\nBackend **FastAPI** integrado ao banco **MariaDB/MySQL** (`miau_db`).\n\n### Autenticação JWT\n\n1. `POST /auth/login` — envie `username` e `password` (form-urlencoded)\n2. Copie o `access_token` retornado\n3. Clique em **Authorize** e informe: `Bearer {token}`\n4. Rotas `/api/*` exigem token válido\n\n**Usuário padrão:** `ShardCadu` / `cadu123`\n\n### Tabelas do banco\n\n| Tabela | Endpoints |\n|--------|-----------|\n| `usuarios` | `/auth/login`, `/auth/register`, `/auth/me` |\n| `tutores` | `/api/tutores` |\n| `pets` | `/api/pets` |\n| `servicos` | `/api/servicos` |\n| `agendamentos` | `/api/agendamentos` |\n| `avisos` | `/api/avisos` |\n\nDocumentação completa: repositório → `docs/API-Swagger.md`\n",
    "version": "1.0.0",
    "contact": {
      "name": "Carlos Eduardo (Cadu)",
      "url": "https://github.com/Dev-Carlos-Alves/MiAu-"
    },
    "license": {
      "name": "Projeto Acadêmico — Henning 2026.1"
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "Local (python app.py)"
    },
    {
      "url": "https://mi-au.vercel.app",
      "description": "Producao Vercel"
    }
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "tags": [
          "Autenticação"
        ],
        "summary": "Registrar novo usuário",
        "description": "Cria registro na tabela `usuarios` e retorna token JWT.",
        "operationId": "register_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Autenticação"
        ],
        "summary": "Login (obter token JWT)",
        "description": "Autentica com username/password (form-urlencoded). Use o token em `Authorization: Bearer`.",
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_login_auth_login_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": [
          "Autenticação"
        ],
        "summary": "Obter perfil autenticado",
        "description": "Retorna username e email do usuário logado. Requer JWT.",
        "operationId": "get_me_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "Autenticação"
        ],
        "summary": "Atualizar perfil autenticado",
        "description": "Atualiza username, email e/ou senha na tabela `usuarios`. Requer JWT.",
        "operationId": "update_me_auth_me_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_login_auth_login_post": {
        "properties": {
          "grant_type": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^password$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grant Type"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "format": "password",
            "title": "Password"
          },
          "scope": {
            "type": "string",
            "title": "Scope",
            "default": ""
          },
          "client_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          },
          "client_secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "format": "password",
            "title": "Client Secret"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "Body_login_auth_login_post"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MessageResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "example": "Registro deletado com sucesso"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "MessageResponse"
      },
...
      "ProfileResponse": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username",
            "example": "ShardCadu"
          },
          "email": {
            "type": "string",
            "title": "Email",
            "example": "cadu.sport@miau.com"
          }
        },
        "type": "object",
        "required": [
          "username",
          "email"
        ],
        "title": "ProfileResponse"
      },
      "ProfileUpdate": {
        "properties": {
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username",
            "example": "ShardCadu"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email",
            "example": "cadu.sport@miau.com"
          },
          "password": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Password",
            "description": "Opcional — deixe vazio para manter",
            "example": "nova_senha"
          }
        },
        "type": "object",
        "title": "ProfileUpdate"
      },
...
      "Token": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token",
            "description": "Token JWT para Authorization Bearer"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer",
            "example": "bearer"
          }
        },
        "type": "object",
        "required": [
          "access_token"
        ],
        "title": "Token"
      },
...
      "UserCreate": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 50,
            "minLength": 3,
            "title": "Username",
            "description": "Nome de usuário único",
            "example": "ShardCadu"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "E-mail único do usuário",
            "example": "cadu.sport@miau.com"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "title": "Password",
            "description": "Senha de acesso",
            "example": "cadu123"
          }
        },
        "type": "object",
        "required": [
          "username",
          "email",
          "password"
        ],
        "title": "UserCreate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/login"
          }
        },
        "description": "Autenticação JWT. No Authorize, use username `ShardCadu` e password `cadu123`. Deixe client_id e client_secret vazios. Rotas /api/* exigem token."
      },
      "BearerJWT": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Alternativa manual: faça POST /auth/login, copie access_token e cole aqui (apenas o token, sem a palavra Bearer)."
      }
    }
  },
...
}